Start of refactoring to use OpenGL ES 2.0
