+++ app/src/main/assets/shaders/circle.vert
@@ -1,0 +1,9 @@
+#version 100
+precision highp float;
+
+attribute vec2 position;
+
+void main() {
+    gl_Position = vec4(position, 0.0, 1.0);
+    gl_PointSize = 300.0;
+}
