+++ app/src/main/java/com/dozingcatsoftware/bouncy/FieldLayout.java
@@ -21,6 +21,7 @@ import android.util.Log;
 import com.dozingcatsoftware.bouncy.elements.FieldElement;
 import com.dozingcatsoftware.bouncy.elements.FieldElementCollection;
 import com.dozingcatsoftware.bouncy.elements.FlipperElement;
+import com.dozingcatsoftware.bouncy.util.IOUtils;
 import com.dozingcatsoftware.bouncy.util.JSONUtils;
 
 public class FieldLayout {
@@ -43,8 +44,7 @@ public class FieldLayout {
     static final String ELEMENTS_PROPERTY = "elements";
 
     static int _numLevels = -1;
-    static Map<Integer, Map<String, Object>> _layoutMap =
-            new HashMap<Integer, Map<String, Object>>();
+    static Map<Integer, Map<String, Object>> _layoutMap = new HashMap<>();
     static Context _context;
     Random RAND = new Random();
 
@@ -75,15 +75,8 @@ public class FieldLayout {
         try {
             String assetPath = "tables/table" + level + ".json";
             InputStream fin = _context.getAssets().open(assetPath);
-            BufferedReader br = new BufferedReader(new InputStreamReader(fin));
-
-            StringBuilder buffer = new StringBuilder();
-            String line;
-            while ((line = br.readLine()) != null) {
-                buffer.append(line);
-            }
-            fin.close();
-            Map<String, Object> layoutMap = JSONUtils.mapFromJSONString(buffer.toString());
+            String s = IOUtils.utf8FromStream(fin);
+            Map<String, Object> layoutMap = JSONUtils.mapFromJSONString(s);
             return layoutMap;
         }
         catch (Exception ex) {
